services/common/pyproject.toml
[build-system]
requires = ["setuptools==82.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "exocort-services-common"
version = "0.1.0"
description = "Shared utilities for exocort services"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.135.1",
  "pydantic==2.12.5",
  "python-dotenv==1.2.2",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["common*"]
